Created a working model
I went through several different API solutions in the interum to replace the failed gtranslate API. Eventually I settled on a Google Scripts webapp that proxys requests from Google Translate. I've adapted the implementation of my programs to work with the new Translation class, which acts as a gateway between the aforementioned webapp and my code. Additionally, I ran through my code and added documentation to the places I hadn't earlier. This is currently a working model, but has none of the tests or deliverables included